+++ app/components/contents/ContentComicItemV2.tsx
@@ -1,42 +1,13 @@
 "use client"
-import { EStorageType } from '@/app/models/enums/EStorageType';
-import { generateImageUrlByStorageType } from '@/app/utils/HelperFunctions';
-import { decryptUrl } from '@/lib/security/securityHelper';
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import LazyLoad from 'react-lazyload';
 
-function ContentComicItemPlaceHolder() {
-    return (
-        <img
-            src={generateImageUrlByStorageType(EStorageType.S1, "1.1_abcda.jpg")}
-            height={1000}
-            alt=""
-            width={800}
-        />
-    );
-}
-
-export function ContentComicItemV2({ imageUrl, storageType }: { imageUrl: string, storageType: EStorageType }) {
-    const [originUrl, setOriginUrl] = useState<string>('');
-
-    useEffect(() => {
-        const fileName = decryptUrl(imageUrl);
-        const url = generateImageUrlByStorageType(storageType, fileName);
-        setOriginUrl(url);
-    }, []);
-
-    const onError = () => {
-        setOriginUrl("/assets/media/404/1.1_abcda.jpg");
-    }
-
+export function ContentComicItemV2({ imageUrl }: { imageUrl: string }) {
     return (
         <div className="chapter-image col-lg-10 offset-lg-1 col-12 offset-0 img-chapter">
-            <LazyLoad height={1000} once={false} offset={1600}
-                placeholder={<ContentComicItemPlaceHolder />}
-                unmountIfInvisible={true}>
+            <LazyLoad height={1000} once={true} offset={1600}>
                 <img
-                    src={originUrl}
-                    onError={onError}
+                    src={imageUrl}
                     alt=""
                     width={800}
                 />
